requirements.txt
openai-whisper
sounddevice
scipy
numpy
requests
kokoro-onnx
soundfile
pyaudio
huggingface_hub
python-dotenv
playwright
playwright-stealth
pyautogui
pillow
psutil
pygetwindow
easyocr
opencv-python
librosa
torch
torchvision
torchaudio
apscheduler
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
spotipy
python-dateutil
